updated sink interface: removed queue and shutdown funcs, updated sinkI update func, added send
