+++ sink.go
@@ -1,112 +1,177 @@
 package main
 
 import (
+	"bytes"
+	"container/ring"
+	"crypto/hmac"
+	"crypto/sha1"
+	"encoding/hex"
 	"fmt"
+	"io"
+	"net/http"
+	"net/url"
+	"strconv"
+	"strings"
 	"time"
 
+	"github.com/xmidt-org/retry"
+	"github.com/xmidt-org/retry/retryhttp"
 	"github.com/xmidt-org/wrp-go/v3"
+	"github.com/xmidt-org/wrp-go/v3/wrphttp"
 	"go.uber.org/zap"
 )
 
 type SinkI interface {
 	Update(Listener) error
-	Shutdown(bool)
-	Queue(*wrp.Message)
+	Send(*ring.Ring, string, string, *wrp.Message) error
 }
 
-type Sink struct {
+type SinkWebhookV1 struct {
+	id               string
+	deliveryInterval time.Duration
+	deliveryRetries  int
+	logger           *zap.Logger
 }
 
-func (s *Sink) Update(l Listener) (err error) {
-	switch v := l.(type) {
-	case *ListenerV1:
-		m := &MatcherV1{}
-		if err = m.Update(*v); err != nil {
-			return
-		}
-		s.matcher = m
-	default:
-		err = fmt.Errorf("invalid listner")
+func NewSinkWebhookV1(s *SinkSender) {
+	v1 := &SinkWebhookV1{
+		id:               s.id,
+		deliveryInterval: s.deliveryInterval,
+		deliveryRetries:  s.deliveryRetries,
 	}
-
-	s.renewalTimeGauge.Set(float64(time.Now().Unix()))
-
-	s.mutex.Lock()
-	defer s.mutex.Unlock()
-
-	s.deliverUntil = l.GetUntil()
-	s.deliverUntilGauge.Set(float64(s.deliverUntil.Unix()))
-	s.deliveryRetryMaxGauge.Set(float64(s.deliveryRetries))
-
-	s.id = l.GetId()
-	s.listener = l
-	s.failureMessage.Original = l
-
-	// Update this here in case we make this configurable later
-	s.maxWorkersGauge.Set(float64(s.maxWorkers))
-
-	return
+	s.sink = v1
+}
+func (v1 *SinkWebhookV1) Update(l Listener) (err error) {
+	v1.id = l.GetId()
+	return nil
 }
 
-// Shutdown causes the CaduceusOutboundSender to stop its activities either gently or
-// abruptly based on the gentle parameter.  If gentle is false, all queued
-// messages will be dropped without an attempt to send made.
-func (s *Sink) Shutdown(gentle bool) {
-	if !gentle {
-		// need to close the channel we're going to replace, in case it doesn't
-		// have any events in it.
-		close(s.queue.Load().(chan *wrp.Message))
-		s.Empty(s.droppedExpiredCounter)
-	}
-	close(s.queue.Load().(chan *wrp.Message))
-	s.wg.Wait()
+// worker is the routine that actually takes the queued messages and delivers
+// them to the listeners outside webpa
+func (v1 *SinkWebhookV1) Send(urls *ring.Ring, secret, acceptType string, msg *wrp.Message) error {
+	defer func() {
+		if r := recover(); nil != r {
+			// s.droppedPanic.Add(1.0)
+			v1.logger.Error("goroutine send() panicked", zap.String("id", v1.id), zap.Any("panic", r))
+		}
+		// s.workers.Release()
+		// s.currentWorkersGauge.Add(-1.0)
+	}()
 
-	s.mutex.Lock()
-	s.deliverUntil = time.Time{}
-	s.deliverUntilGauge.Set(float64(s.deliverUntil.Unix()))
-	s.queueDepthGauge.Set(0) //just in case
-	s.mutex.Unlock()
+	payload := msg.Payload
+	body := payload
+	var payloadReader *bytes.Reader
+
+	// Use the internal content type unless the accept type is wrp
+	contentType := msg.ContentType
+	switch acceptType {
+	case "wrp", wrp.MimeTypeMsgpack, wrp.MimeTypeWrp:
+		// WTS - We should pass the original, raw WRP event instead of
+		// re-encoding it.
+		contentType = wrp.MimeTypeMsgpack
+		buffer := bytes.NewBuffer([]byte{})
+		encoder := wrp.NewEncoder(buffer, wrp.Msgpack)
+		encoder.Encode(msg)
+		body = buffer.Bytes()
+	}
+	payloadReader = bytes.NewReader(body)
+
+	req, err := http.NewRequest("POST", urls.Value.(string), payloadReader)
+	if err != nil {
+		// Report drop
+		// s.droppedInvalidConfig.Add(1.0)
+		v1.logger.Error("Invalid URL", zap.String("url", urls.Value.(string)), zap.String("id", v1.id), zap.Error(err))
+		return err
+	}
+
+	req.Header.Set("Content-Type", contentType)
+
+	// Add x-Midt-* headers
+	wrphttp.AddMessageHeaders(req.Header, msg)
+
+	// Provide the old headers for now
+	req.Header.Set("X-Webpa-Event", strings.TrimPrefix(msg.Destination, "event:"))
+	req.Header.Set("X-Webpa-Transaction-Id", msg.TransactionUUID)
+
+	// Add the device id without the trailing service
+	id, _ := wrp.ParseDeviceID(msg.Source)
+	req.Header.Set("X-Webpa-Device-Id", string(id))
+	req.Header.Set("X-Webpa-Device-Name", string(id))
+
+	// Apply the secret
+
+	if secret != "" {
+		s := hmac.New(sha1.New, []byte(secret))
+		s.Write(body)
+		sig := fmt.Sprintf("sha1=%s", hex.EncodeToString(s.Sum(nil)))
+		req.Header.Set("X-Webpa-Signature", sig)
+	}
+
+	// find the event "short name"
+	event := msg.FindEventStringSubMatch()
+
+	// Send it
+	v1.logger.Debug("attempting to send event", zap.String("event.source", msg.Source), zap.String("event.destination", msg.Destination))
+	client, _ := retryhttp.NewClient(
+		// retryhttp.WithHTTPClient(s.clientMiddleware(s.client)),
+		retryhttp.WithRunner(v1.addRunner(req, event)),
+		retryhttp.WithRequesters(v1.updateRequest(urls)),
+	)
+	resp, err := client.Do(req)
+
+	code := "failure"
+	logger := v1.logger
+	if err != nil {
+		// Report failure
+		// s.droppedNetworkErrCounter.Add(1.0)
+		logger.Error("Dropped Network Error", zap.Error(err))
+		return err
+	} else {
+		// Report Result
+		code = strconv.Itoa(resp.StatusCode)
+
+		// read until the response is complete before closing to allow
+		// connection reuse
+		if resp.Body != nil {
+			io.Copy(io.Discard, resp.Body)
+			resp.Body.Close()
+		}
+	}
+	// s.deliveryCounter.With(prometheus.Labels{UrlLabel: s.id, CodeLabel: code, EventLabel: event}).Add(1.0)
+	logger.Debug("event sent-ish", zap.String("event.source", msg.Source), zap.String("event.destination", msg.Destination), zap.String("code", code), zap.String("url", req.URL.String()))
+	return nil
 }
 
-// Queue is given a request to evaluate and optionally enqueue in the list
-// of messages to deliver.  The request is checked to see if it matches the
-// criteria before being accepted or silently dropped.
-// TODO: can pass in message along with webhook information
-func (s *Sink) Queue(msg *wrp.Message) {
-	s.mutex.RLock()
-	deliverUntil := s.deliverUntil
-	dropUntil := s.dropUntil
-	s.mutex.RUnlock()
+func (v1 *SinkWebhookV1) addRunner(request *http.Request, event string) retry.Runner[*http.Response] {
+	runner, _ := retry.NewRunner[*http.Response](
+		retry.WithPolicyFactory[*http.Response](retry.Config{
+			Interval:   v1.deliveryInterval,
+			MaxRetries: v1.deliveryRetries,
+		}),
+		retry.WithOnAttempt[*http.Response](v1.onAttempt(request, event)),
+	)
+	return runner
+}
 
-	now := time.Now()
-
-	if !s.isValidTimeWindow(now, dropUntil, deliverUntil) {
-		s.logger.Debug("invalid time window for event", zap.Any("now", now), zap.Any("dropUntil", dropUntil), zap.Any("deliverUntil", deliverUntil))
-		return
+func (v1 *SinkWebhookV1) updateRequest(urls *ring.Ring) func(*http.Request) *http.Request {
+	return func(request *http.Request) *http.Request {
+		urls = urls.Next()
+		tmp, err := url.Parse(urls.Value.(string))
+		if err != nil {
+			v1.logger.Error("failed to update url", zap.String(UrlLabel, urls.Value.(string)), zap.Error(err))
+		}
+		request.URL = tmp
+		return request
 	}
+}
 
-	if !s.disablePartnerIDs {
-		if len(msg.PartnerIDs) == 0 {
-			msg.PartnerIDs = s.customPIDs
-		}
-		if !overlaps(s.listener.GetPartnerIds(), msg.PartnerIDs) {
-			s.logger.Debug("parter id check failed", zap.Strings("webhook.partnerIDs", s.listener.GetPartnerIds()), zap.Strings("event.partnerIDs", msg.PartnerIDs))
-			return
-		}
-		if ok := s.matcher.IsMatch(msg); !ok {
-			return
+func (v1 *SinkWebhookV1) onAttempt(request *http.Request, event string) retry.OnAttempt[*http.Response] {
+
+	return func(attempt retry.Attempt[*http.Response]) {
+		if attempt.Retries > 0 {
+			// s.deliveryRetryCounter.With(prometheus.Labels{UrlLabel: v1.id, EventLabel: event}).Add(1.0)
+			v1.logger.Debug("retrying HTTP transaction", zap.String("url", request.URL.String()), zap.Error(attempt.Err), zap.Int("retry", attempt.Retries+1), zap.Int("statusCode", attempt.Result.StatusCode))
 		}
 
-	}
-
-	//TODO: this code will be changing will take away queue and send to the sink interface (not webhook interface)
-	select {
-	case s.queue.Load().(chan *wrp.Message) <- msg:
-		s.queueDepthGauge.Add(1.0)
-		s.logger.Debug("event added to outbound queue", zap.String("event.source", msg.Source), zap.String("event.destination", msg.Destination))
-	default:
-		s.logger.Debug("queue full. event dropped", zap.String("event.source", msg.Source), zap.String("event.destination", msg.Destination))
-		s.queueOverflow()
-		s.droppedQueueFullCounter.Add(1.0)
 	}
 }
