+++ sinkSender.go
@@ -13,9 +13,6 @@ import (
 	"fmt"
 	"io"
 	"net/http"
-	"net/url"
-	"strconv"
-	"strings"
 	"sync"
 	"sync/atomic"
 	"time"
@@ -23,11 +20,8 @@ import (
 	"go.uber.org/zap"
 
 	"github.com/prometheus/client_golang/prometheus"
-	"github.com/xmidt-org/retry"
-	"github.com/xmidt-org/retry/retryhttp"
 	"github.com/xmidt-org/webpa-common/v2/semaphore"
 	"github.com/xmidt-org/wrp-go/v3"
-	"github.com/xmidt-org/wrp-go/v3/wrphttp"
 )
 
 // failureText is human readable text for the failure message
@@ -70,6 +64,7 @@ type SinkSender struct {
 	workers           semaphore.Interface
 	logger            *zap.Logger
 	sink              SinkI
+	client            http.Client
 	clientMiddleware  func(Client) Client
 	failureMessage    FailureMessage
 	listener          Listener
@@ -139,6 +134,7 @@ func NewSinkSender(sw *SinkWrapper, l Listener) (sender *SinkSender, err error) 
 		disablePartnerIDs: sw.config.DisablePartnerIDs,
 		clientMiddleware:  sw.clientMiddleware,
 	}
+
 	sender.CreateMetrics(sw.metrics)
 	sender.queueDepthGauge.Set(0)
 	sender.currentWorkersGauge.Set(0)
@@ -167,6 +163,7 @@ func (s *SinkSender) Update(l Listener) (err error) {
 			return
 		}
 		s.matcher = m
+		NewSinkWebhookV1(s)
 	default:
 		err = fmt.Errorf("invalid listner")
 	}
@@ -440,141 +437,11 @@ Loop:
 			s.workers.Acquire()
 			s.currentWorkersGauge.Add(1.0)
 
-			go s.send(urls, secret, accept, msg)
+			go s.sink.Send(urls, secret, accept, msg)
 		}
 	}
 	for i := 0; i < s.maxWorkers; i++ {
 		s.workers.Acquire()
-	}
-}
-
-//TODO:
-
-// worker is the routine that actually takes the queued messages and delivers
-// them to the listeners outside webpa
-func (s *SinkSender) send(urls *ring.Ring, secret, acceptType string, msg *wrp.Message) {
-	defer func() {
-		if r := recover(); nil != r {
-			s.droppedPanic.Add(1.0)
-			s.logger.Error("goroutine send() panicked", zap.String("id", s.id), zap.Any("panic", r))
-		}
-		s.workers.Release()
-		s.currentWorkersGauge.Add(-1.0)
-	}()
-
-	payload := msg.Payload
-	body := payload
-	var payloadReader *bytes.Reader
-
-	// Use the internal content type unless the accept type is wrp
-	contentType := msg.ContentType
-	switch acceptType {
-	case "wrp", wrp.MimeTypeMsgpack, wrp.MimeTypeWrp:
-		// WTS - We should pass the original, raw WRP event instead of
-		// re-encoding it.
-		contentType = wrp.MimeTypeMsgpack
-		buffer := bytes.NewBuffer([]byte{})
-		encoder := wrp.NewEncoder(buffer, wrp.Msgpack)
-		encoder.Encode(msg)
-		body = buffer.Bytes()
-	}
-	payloadReader = bytes.NewReader(body)
-
-	req, err := http.NewRequest("POST", urls.Value.(string), payloadReader)
-	if nil != err {
-		// Report drop
-		s.droppedInvalidConfig.Add(1.0)
-		s.logger.Error("Invalid URL", zap.String("url", urls.Value.(string)), zap.String("id", s.id), zap.Error(err))
-		return
-	}
-
-	req.Header.Set("Content-Type", contentType)
-
-	// Add x-Midt-* headers
-	wrphttp.AddMessageHeaders(req.Header, msg)
-
-	// Provide the old headers for now
-	req.Header.Set("X-Webpa-Event", strings.TrimPrefix(msg.Destination, "event:"))
-	req.Header.Set("X-Webpa-Transaction-Id", msg.TransactionUUID)
-
-	// Add the device id without the trailing service
-	id, _ := wrp.ParseDeviceID(msg.Source)
-	req.Header.Set("X-Webpa-Device-Id", string(id))
-	req.Header.Set("X-Webpa-Device-Name", string(id))
-
-	// Apply the secret
-
-	if secret != "" {
-		s := hmac.New(sha1.New, []byte(secret))
-		s.Write(body)
-		sig := fmt.Sprintf("sha1=%s", hex.EncodeToString(s.Sum(nil)))
-		req.Header.Set("X-Webpa-Signature", sig)
-	}
-
-	// find the event "short name"
-	event := msg.FindEventStringSubMatch()
-
-	// Send it
-	s.logger.Debug("attempting to send event", zap.String("event.source", msg.Source), zap.String("event.destination", msg.Destination))
-	client, _ := retryhttp.NewClient(
-		// retryhttp.WithHTTPClient(s.clientMiddleware(s.client)),
-		retryhttp.WithRunner(s.addRunner(req, event)),
-		retryhttp.WithRequesters(s.updateRequest(urls)),
-	)
-	resp, err := client.Do(req)
-
-	code := "failure"
-	logger := s.logger
-	if nil != err {
-		// Report failure
-		s.droppedNetworkErrCounter.Add(1.0)
-		logger = s.logger.With(zap.Error(err))
-	} else {
-		// Report Result
-		code = strconv.Itoa(resp.StatusCode)
-
-		// read until the response is complete before closing to allow
-		// connection reuse
-		if nil != resp.Body {
-			io.Copy(io.Discard, resp.Body)
-			resp.Body.Close()
-		}
-	}
-	s.deliveryCounter.With(prometheus.Labels{UrlLabel: s.id, CodeLabel: code, EventLabel: event}).Add(1.0)
-	logger.Debug("event sent-ish", zap.String("event.source", msg.Source), zap.String("event.destination", msg.Destination), zap.String("code", code), zap.String("url", req.URL.String()))
-}
-
-func (s *SinkSender) addRunner(request *http.Request, event string) retry.Runner[*http.Response] {
-	runner, _ := retry.NewRunner[*http.Response](
-		retry.WithPolicyFactory[*http.Response](retry.Config{
-			Interval:   s.deliveryInterval,
-			MaxRetries: s.deliveryRetries,
-		}),
-		retry.WithOnAttempt[*http.Response](s.onAttempt(request, event)),
-	)
-	return runner
-}
-
-func (s *SinkSender) updateRequest(urls *ring.Ring) func(*http.Request) *http.Request {
-	return func(request *http.Request) *http.Request {
-		urls = urls.Next()
-		tmp, err := url.Parse(urls.Value.(string))
-		if err != nil {
-			s.logger.Error("failed to update url", zap.String(UrlLabel, urls.Value.(string)), zap.Error(err))
-		}
-		request.URL = tmp
-		return request
-	}
-}
-
-func (s *SinkSender) onAttempt(request *http.Request, event string) retry.OnAttempt[*http.Response] {
-
-	return func(attempt retry.Attempt[*http.Response]) {
-		if attempt.Retries > 0 {
-			s.deliveryRetryCounter.With(prometheus.Labels{UrlLabel: s.id, EventLabel: event}).Add(1.0)
-			s.logger.Debug("retrying HTTP transaction", zap.String("url", request.URL.String()), zap.Error(attempt.Err), zap.Int("retry", attempt.Retries+1), zap.Int("statusCode", attempt.Result.StatusCode))
-		}
-
 	}
 }
